[Decomp][ScriptMgr] Cohesion split: extract On* scripting-hook dispatchers to ScriptMgrHooks.cpp
Move the scripting-hook forwarders out of ScriptMgr.cpp into a dedicated TU
(verbatim move, no behaviour change; ScriptMgr.h unchanged, CMake
file(GLOB WorldHandlers/*.cpp) auto-registers it):

  ScriptMgrHooks.cpp  GetCreatureAI/GetGameObjectAI/CreateInstanceData factories
                      + the On{GossipHello,GossipSelect,QuestAccept,QuestRewarded,
                      GameObjectUse,ItemUse,AreaTrigger,NpcSpellClick,ProcessEvent,
                      EffectDummy,EffectScriptEffect,AuraDummy} + GetDialogStatus
                      dispatchers forwarding to SD3 and Eluna.

Completes the ScriptMgr.cpp split (follow-up to #257): the ctor/dtor and the
LoadScriptLibrary/UnloadScriptLibrary/CollectPossibleEventIds core remain.
ScriptMgr.cpp 1,046 -> 345.
